hw_access: Update Scom access for new system device tree format
The new system device tree has static paths for targets which get
modelled according to the actual hardware layout. This means that a
processor module (PROC_CHIP) gets modelled with the various busses
underneath it. Therefore when doing a SCOM access on a PROC_CHIP we
need to lookup the actual pib bus target first to pass to the pdbg
access function.

This patch implements the lookup by providing a specific
specialisation for PROC_CHIP SCOM access methods.

Signed-off-by: Alistair Popple <alistair@popple.id.au>