hwpf: fix crosses initialization error in switch case
Error parser need update to support , error xml which got
variable initlisation inside case statement. Existing
parser script is failing due to this

The C++ standard says: "It is possible to transfer into
a block, but not in a way that bypasses declarations with
initialization. A program that jumps from a point where
a local variable with automatic storage duration is not
in scope to a point where it is in scope is ill-formed
unless the variable has POD type (3.9) and is declared
without an initializer."

Fix is to update parser script to add enclose each case
in its own local scope creates.

Signed-off-by: Jayanth Othayoth <ojayanth@in.ibm.com>