configure: Install attribute database with --enable-db-install
The attribute database was added to the list of files to install when
not building for arm. This was an attempt to detect the BMC, however
future BMCs may not match this pattern so instead control installation
with a new configure time argument.

The database is always built as it is used to generate platform_attrs.h.

Signed-off-by: Joel Stanley <joel@jms.id.au>